# strava_analytics/web/openapi.yaml
openapi: "3.1.0"
info:
  title: MERON API
  description: >
    Personal training data API. Provides fitness metrics, activity CRUD,
    training plans, race predictions, and Strava sync. All distances in
    miles, weights in lbs, elevations in feet. Times in seconds unless
    otherwise noted.
  version: "2.0.0"
servers:
  - url: http://localhost:8051
    description: Local standalone API server (strava-api)
  - url: http://localhost:8050
    description: Local Dash + API combined server (strava-web)
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Sign in with username + password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: Session cookie set
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/SessionUser" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/auth/me:
    get:
      tags: [auth]
      operationId: me
      summary: Current session info (401 when not logged in)
      security: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionUser" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/signup:
    post:
      tags: [auth]
      operationId: signup
      summary: Create an account using an invite code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, invite_code]
              properties:
                username: { type: string, minLength: 3, maxLength: 64 }
                password: { type: string, minLength: 8, format: password }
                invite_code: { type: string }
      responses:
        "201":
          description: Created and logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/SessionUser" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/auth/invites:
    get:
      tags: [auth]
      operationId: listInvites
      summary: Admin — list outstanding invite codes
      responses:
        "200":
          description: Invite list
          content:
            application/json:
              schema:
                type: object
                properties:
                  invites:
                    type: array
                    items:
                      type: object
                      properties:
                        code: { type: string }
                        created_at: { type: string, format: date-time }
                        expires_at:
                          type: string
                          format: date-time
                          nullable: true
    post:
      tags: [auth]
      operationId: createInvite
      summary: Admin — generate a new invite code
      responses:
        "201":
          description: New code
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/invites/{code}:
    parameters:
      - name: code
        in: path
        required: true
        schema: { type: string }
    delete:
      tags: [auth]
      operationId: revokeInvite
      summary: Admin — delete an unused invite code
      responses:
        "200":
          description: Revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/auth/pair/claim:
    post:
      tags: [auth]
      operationId: claimPairCode
      summary: Mobile app bootstrap — exchange a pair code for an API key
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  description: 8-character pair code generated in Settings
      responses:
        "200":
          description: Claimed — stash api_key + api_base on the device
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_key: { type: string }
                  api_base: { type: string, format: uri }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Read key for GET; write key for POST/PATCH/DELETE.
  responses:
    BadRequest:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: object
              additionalProperties: true
          required: [code, message]
...
    SessionUser:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        display_name: { type: string, nullable: true }
        is_admin: { type: boolean }
...